various changes

+++ src/components/common/apiCalls.js
@@ -1,6 +1,7 @@
 import format from 'date-fns/format';
 import url from '../urls';
 import {reduceNames} from "../../dev/helpers";
+import {reduceSessions, reduceSessionsByTimePeriod} from "../../utils/reduceSessions";
 
 const env = 'dev';
 const urls = url(env);
@@ -11,13 +12,33 @@ export const getSessions = async (date, timePeriod = 'day') => {
     if (env === 'dev') {
         if (timePeriod === 'day') {
             response = await fetch(`${urls.sessions}?date=${date}`);
+            if (response.status >= 400) {
+                throw(new Error('Error fetching sessions'))
+            } else {
+                return await response.json()
+            }
+        } else if (timePeriod === 'allTime') {
+            //fetch all sessions by name
+            response = await fetch(`${urls.sessions}`);
+            if (response.status >= 400) {
+                throw(new Error('Error fetching sessions'))
+            } else {
+                const sessions = await response.json();
+                return new Promise((resolve) => {
+                    resolve(reduceSessions(sessions));
+                });
+            }
         } else {
+            //fetch sessions in year or month
             response = await fetch(`${urls.sessions}?q=${date}`);
-        }
-        if (response.status >= 400) {
-            throw(new Error('Error fetching sessions'))
-        } else {
-            return await response.json()
+            if (response.status >= 400) {
+                throw(new Error('Error fetching sessions'))
+            } else {
+                const sessions = await response.json();
+                return new Promise((resolve) => {
+                    resolve(reduceSessionsByTimePeriod(sessions, timePeriod));
+                });
+            }
         }
     }
 };
@@ -33,35 +54,74 @@ export const getSessions = async (date, timePeriod = 'day') => {
 
 //PATCH session (update)
 export const patchSessions = async (id, sessionName) => {
-    const init = {
-        method: "PATCH",
-        body: JSON.stringify({sessionName}),
-        headers: {
-            'Content-Type': 'application/json'
+    if (env === 'dev') {
+        // id exists so patch single session
+        if (id !== undefined) {
+            const init = {
+                method: "PATCH",
+                body: JSON.stringify({sessionName}),
+                headers: {
+                    'Content-Type': 'application/json'
+                }
+            };
+            const response = await fetch(`${urls.sessions}/${id}`, init);
+            if (response.status >= 400) {
+                throw(new Error('Error update sessions'))
+            } else {
+                return await response.json()
+            }
+        } else if (id === undefined) {
+            const init = {
+                method: "PATCH",
+                body: JSON.stringify({sessionName}),
+                headers: {
+                    'Content-Type': 'application/json'
+                }
+            };
+
+            const response = await fetch(`${urls.sessions}/?sessionName=${sessionName}`, init);
+            if (response.status >= 400) {
+                throw(new Error('Error update sessions'))
+            } else {
+                return await response.json()
+            }
         }
-    };
-    const response = await fetch(`${urls.sessions}/${id}`, init);
-    if (response.status >= 400) {
-        throw(new Error('Error update sessions'))
-    } else {
-        return await response.json()
     }
 };
 
 //DELETE session(delete)
 export const deleteSessions = async (id) => {
-    const init = {
-        method: "DELETE",
-        headers: {
-            'Content-Type': 'application/json'
+    if (env === 'dev') {
+        // id exists so patch single session
+        if (id !== undefined) {
+            const init = {
+                method: "DELETE",
+                headers: {
+                    'Content-Type': 'application/json'
+                }
+            };
+            const response = await fetch(`${urls.sessions}/${id}`, init);
+            if (response.status >= 400) {
+                throw(new Error('Error in delete sessions'))
+            } else {
+                return await response.json()
+            }
+        } else if (id === undefined) {
+            const init = {
+                method: "DELETE",
+                headers: {
+                    'Content-Type': 'application/json'
+                }
+            };
+            const response = await fetch(`${urls.sessions}/?sessionName=gatis`, init);
+            if (response.status >= 400) {
+                throw(new Error('Error in delete sessions'))
+            } else {
+                return await response.json()
+            }
         }
-    };
-    const response = await fetch(`${urls.sessions}/${id}`, init);
-    if (response.status >= 400) {
-        throw(new Error('Error in delete sessions'))
-    } else {
-        return await response.json()
     }
+
 };
 
 export const postSessions = async (sessionName, duration) => {
@@ -85,12 +145,18 @@ export const postSessions = async (sessionName, duration) => {
     }
 };
 
-export const getState = async () => {
-    const response = await fetch(urls.state);
-    if (response.status >= 400) {
-        throw(new Error('Error fetching state'))
-    } else {
-        return await response.json()
+export const getSessionState = async () => {
+    if (env === 'dev') {
+        const responseState = await fetch(urls.state);
+        // const minDate = await getMinDate();
+        if (responseState.status >= 400) {
+            throw(new Error('Error fetching state'))
+        } else {
+            return await responseState.json();
+            // return new Promise((resolve) => {
+            //     resolve({sessionState, appState: {minDate}});
+            // });
+        }
     }
 };
 
@@ -134,42 +200,17 @@ export const getNames = async (count) => {
     }
 };
 
-// export const getMinDate = async () => {
-//     const response = await fetch(urls.total);
-//     if (response.status >= 400) {
-//         throw(new Error('Error fetching minDate'))
-//     } else {
-//         return await response.json()
-//     }
-// };
-
-// this is for dev server fetches
-export const getSessionsDurations = async (date, length) => {
-    let response;
-    if (length === 'month') {
-        response = await fetch(`${urls.sessionsMonth}?timePeriod=${date}`);
-    } else if (length === 'year') {
-        response = await fetch(`${urls.sessionsYear}?timePeriod=${date}`);
-    }
-    if (response.status >= 400) {
-        throw(new Error(`Error fetching ${length} sessions`))
-    } else {
-        return await response.json()
+// get minDate
+export const getMinDate = async () => {
+    //development server url
+    if (env === 'dev') {
+        const response = await fetch(urls.sessions);
+        if (response.status >= 400) {
+            throw(new Error('Error fetching minDate'))
+        } else {
+            const sessions = await response.json();
+            //helper function to imitate data returned from server
+            return sessions[0].date;
+        }
     }
 };
-
-// // get mindate
-// export const getMinDate = async () => {
-//     //development server url
-//     if (env === 'dev') {
-//         const response = await fetch(urls.minDate);
-//         if (response.status >= 400) {
-//             throw(new Error('Error fetching minDate'))
-//         } else {
-//             const names = await response.json();
-//             //helper function to imitate data returned from server
-//             return reduceNames(names, count);
-//         }
-//     }
-//     await fetch();
-// };
